enable start/stop buttons based on actual bot running state
This improves these cases, where previously the buttons
would be out-of-sync with the bot state:
 - the page is reloaded while the bot is running
 - the bot exits on its own

This still doesn't enable or disable the input buttons
properly, but it's a start.

Also, remove the manual ALI argument to always connect
to the local bot.  This is isn't necessary as it's the
default anyway, and was making testing harder.