add initial transcript view
Adds an "audio" tab.  This will only be shown if audio
is configured in the oobabot config file and discrivener
is installed.

The transcript view is an html chat view which is updated
live as the audio is transcribed.

It still needs more polish, but I wanted to get it
in for safekeeping.