partial fix for #17: enable streaming setting
The streaming setting couldn't be applied in the web UI, even when
manually editing the yaml, because the option was hidden.  Now that
we have better configuration options for the streaming setting, just
enable the option to make this work.

Also, move the audio tab to the end of the tab order, if it exists.