fix for #17: wakewords lost when using .json persona
When using a json persona, user-supplied wakewords weren't saved.
This was pretty broken!

The mess was with the logic to automatically add the character's
name as a wakeword, which was dropping user-supplied wakewords in
some cases.  I decided to just make this all simpler and let the
wakewords be wakewords, and trust that the bot user can figure
out themselves what they want to enter there.