performance optimization for log viewing
instead of updating the full html log, have a hidden
field that changes each time a new log entry is added.
And then watch this for changes, and only update the
html view when we know a new log entry has been added.

After this change, you will need to run this against the
'head' version of `oobabot`, not the 0.2.0 release.