fix issue 24 -- crash when loading model with new transformers version
Newer versions of the transformers library monkey-patch the
global Logging object, which gets undone when we load.

For now, just manually re-patch the global Logging object
with their monkey patches when we load, using only the
specific attributes that they currently add.  This would
need to be updated if they add more attributes, but is
safer and quicker than doing something more generic at this point.