revert gradio to version 3.33.1
Newer builds of text-generation-webui do not work with
gradio 3.34.0.  This is very unfortunate as this version
of gradio fixes two security vulnerabilities which are present
in earlier versions.  See:
https://github.com/gradio-app/gradio/security/advisories/GHSA-3qqg-pgqq-3695

For this reason we had been depending on 3.34.0, since that
would at least get our users the security fix.  But now that
text-generation-webui will fail to load models with the newer
version, we don't have a choice to but to depend on the same
broken version until they fix the issue.